Move mst.py into networkx.algorithms.tree.
